Ubuntu 16.06 is using systemd
The recipe uses upstart if the platform is ubuntu and the version
is superior or equal to 13.10.
Ubuntu 16.04 is using systemd, so chef fails to enable the service.

This commit moved the init system selection to the attributes level
in order to be able to override the init system in a wrapper cookbook.
By default the attribute is set to nil, chef will determine wich init
system to use.
If the platform is ubuntu and the version is between 13.10 and 16.04
(not included). the init system is forced to upstart.